#328: Moved the item-related mappers to mappers/items
-
