+++ src/app/ffbe/item-recipes/item-recipes.component.ts
@@ -3,7 +3,7 @@ import {FormControl} from '@angular/forms';
 import {ItemRecipesService} from '../services/item-recipes.service';
 import {Recette} from '../model/recette.model';
 import {ItemRecipe} from '../model/items/item-recipe.model';
-import {ItemRecipeMapper} from '../mappers/item-recipe-mapper';
+import {ItemRecipeMapper} from '../mappers/items/item-recipe-mapper';
 import {FfbeUtils} from '../utils/ffbe-utils';
 
 @Component({
