+++ src/app/ffbe/mappers/items/consumable-mapper.spec.ts
@@ -1,7 +1,7 @@
-import {Objet} from '../model/objet/objet.model';
-import {FfbeUtils} from '../utils/ffbe-utils';
-import {CONSUMABLES_TEST_DATA} from '../model/items/consumable/consumable.model.spec';
-import {Consumable} from '../model/items/consumable/consumable.model';
+import {Objet} from '../../model/objet/objet.model';
+import {FfbeUtils} from '../../utils/ffbe-utils';
+import {CONSUMABLES_TEST_DATA} from '../../model/items/consumable/consumable.model.spec';
+import {Consumable} from '../../model/items/consumable/consumable.model';
 import {ConsumableMapper} from './consumable-mapper';
 
 describe('ConsumableMapper', () => {
