+++ src/app/ffbe/mappers/items/consumable-mapper.ts
@@ -1,12 +1,12 @@
-import {Consumable} from '../model/items/consumable/consumable.model';
-import {Objet} from '../model/objet/objet.model';
-import {Caracteristiques} from '../model/caracteristiques.model';
-import {ResistancesElementaires} from '../model/resistances-elementaires.model';
-import {FFBE_ENGLISH_TABLE_INDEX, FFBE_FRENCH_TABLE_INDEX} from '../ffbe.constants';
-import {ItemCategoryFactory} from '../model/items/item-category.model';
-import {CategorieObjet} from '../model/objet/categorie-objet.model';
-import {FfbeUtils} from '../utils/ffbe-utils';
-import {ResistancesAlterations} from '../model/resistances-alterations.model';
+import {Consumable} from '../../model/items/consumable/consumable.model';
+import {Objet} from '../../model/objet/objet.model';
+import {Caracteristiques} from '../../model/caracteristiques.model';
+import {ResistancesElementaires} from '../../model/resistances-elementaires.model';
+import {FFBE_ENGLISH_TABLE_INDEX, FFBE_FRENCH_TABLE_INDEX} from '../../ffbe.constants';
+import {ItemCategoryFactory} from '../../model/items/item-category.model';
+import {CategorieObjet} from '../../model/objet/categorie-objet.model';
+import {FfbeUtils} from '../../utils/ffbe-utils';
+import {ResistancesAlterations} from '../../model/resistances-alterations.model';
 
 export class ConsumableMapper {
