+++ src/app/ffbe/mappers/items/equipment-mapper.spec.ts
@@ -1,18 +1,18 @@
-import {Objet} from '../model/objet/objet.model';
-import {FfbeUtils} from '../utils/ffbe-utils';
-import {EQUIPMENTS_TEST_DATA} from '../model/items/equipment/equipment.model.spec';
+import {Objet} from '../../model/objet/objet.model';
+import {FfbeUtils} from '../../utils/ffbe-utils';
+import {EQUIPMENTS_TEST_DATA} from '../../model/items/equipment/equipment.model.spec';
 import {EquipmentMapper} from './equipment-mapper';
-import {Equipment} from '../model/items/equipment/equipment.model';
-import {CharactersServiceMock} from '../services/characters.service.spec';
-import {CharactersService} from '../services/characters.service';
-import {CHARACTER_TEST_DATA} from '../model/character/character.model.spec';
-import {Character} from '../model/character/character.model';
-import {Caracteristiques} from '../model/caracteristiques.model';
-import {PASSIVE_SKILLS_TEST_DATA} from '../model/skill.model.spec';
-import {Skill} from '../model/skill.model';
-import {ResistancesElementaires} from '../model/resistances-elementaires.model';
-import {ResistancesAlterations} from '../model/resistances-alterations.model';
-import {Tueurs} from '../model/tueurs.model';
+import {Equipment} from '../../model/items/equipment/equipment.model';
+import {CharactersServiceMock} from '../../services/characters.service.spec';
+import {CharactersService} from '../../services/characters.service';
+import {CHARACTER_TEST_DATA} from '../../model/character/character.model.spec';
+import {Character} from '../../model/character/character.model';
+import {Caracteristiques} from '../../model/caracteristiques.model';
+import {PASSIVE_SKILLS_TEST_DATA} from '../../model/skill.model.spec';
+import {Skill} from '../../model/skill.model';
+import {ResistancesElementaires} from '../../model/resistances-elementaires.model';
+import {ResistancesAlterations} from '../../model/resistances-alterations.model';
+import {Tueurs} from '../../model/tueurs.model';
 
 describe('EquipmentMapper', () => {
   it('should transform equipment raw data into Objet', () => {
