+++ src/app/ffbe/mappers/items/materia-mapper.ts
@@ -1,12 +1,12 @@
-import {Materia} from '../model/items/materia/materia.model';
-import {Objet} from '../model/objet/objet.model';
-import {FFBE_ENGLISH_TABLE_INDEX, FFBE_FRENCH_TABLE_INDEX} from '../ffbe.constants';
-import {ItemCategoryFactory} from '../model/items/item-category.model';
-import {SkillMapper} from './skill-mapper';
-import {FfbeUtils} from '../utils/ffbe-utils';
-import {Caracteristiques} from '../model/caracteristiques.model';
+import {Materia} from '../../model/items/materia/materia.model';
+import {Objet} from '../../model/objet/objet.model';
+import {FFBE_ENGLISH_TABLE_INDEX, FFBE_FRENCH_TABLE_INDEX} from '../../ffbe.constants';
+import {ItemCategoryFactory} from '../../model/items/item-category.model';
+import {SkillMapper} from '../skill-mapper';
+import {FfbeUtils} from '../../utils/ffbe-utils';
+import {Caracteristiques} from '../../model/caracteristiques.model';
 import {ItemWithSkillsMapper} from './item-with-skills-mapper';
-import {TueursMapper} from './tueurs-mapper';
+import {TueursMapper} from '../tueurs-mapper';
 
 export class MateriaMapper extends ItemWithSkillsMapper {
